projects/ai_code_reviewer/pyproject.toml
[project]
name = "ai-code-reviewer"
version = "0.1.0"
description = "Provides helpful code reviews"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dynaconf>=3.2.6",
    "langchain-community>=0.3.10",
    "langchain>=0.3.10",
    "pydantic>=2.10.3",
    "requests>=2.32.3",
    "langchain-ollama>=0.2.1",
    "loguru>=0.7.3",
    "PyGithub>=2.1.1",
    "quart>=0.19.4",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.25.0",
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
]
